Insufficient material fixes
Although two knights cannot FORCE a checkmate, the opponent can
blunder. This position is not a draw.

Also, if each side has one minor piece (unless they are two
Bishops on the same color) the game is - surprisingly - not drawn.
With the extra piece you could wind up in a checkmate that's not
possible if the King is on his own.

https://bugzilla.gnome.org/show_bug.cgi?id=691053