Behave properly when no engines are installed
ai_profiles only contains the list of installed engines, so this can
reasonably be expected to be empty on distros which do not require a
chess engine (such as Arch). Make sure it's not so we don't segfault.

https://bugzilla.gnome.org/show_bug.cgi?id=700330

If no engines are installed, the opponent combo box will be blank, which
is confusing. Moreover the game will break completely in this case if
you set to play Black. So actually set the opponent to "human".