Stop the chess clock when ChessGame is freed
Otherwise, the timer event remains in the main loop, and causes us to
crash when it expires. This would happen if the game is not stopped
before starting a new one, e.g. when the player uses "New Game" or "Load
Game" before the previous game has played to completion.

https://bugzilla.gnome.org/show_bug.cgi?id=701705