Remove the menubar, add an app menu, and expand the toolbar
Not necessarily the final design. All the other games have New Game in
the app menu. But I hesitate to put it there without Open and Save.
These are all very much window-specific options anyway.

Maybe it's OK to have window-specific options there, if the app is
single instance. I'm not sure anyone really knows.

And maybe we want Chess to allow multiple instances again. So not
necessarily the final design.