Remove face-to-face board orientation
This isn't a board orientation, it's a piece orientation.  It actually
silently forces the "White" board orientation.

It could be implemented separately as a checkbox, but I don't think we
want to keep this "feature."

Also, this option is broken in 3D mode.  (Though as a checkbox, it could
be insensitive in 3D mode.)