Use sd to control difficulty of most CECP engines
I was disappointed when I discovered that most engines did not respect
the depth command.  But it turns out that this is a GNUChess (and
Phalanx)-specific extension.  Everyone else uses sd for the same
purpose. This means that we get to provide meaningful difficulties for
these engines after all. Yay!