Handle pause correctly before the game has started
Since the clock doesn't start until after White's first move, it
shouldn't be possible to pause/unpause the game during this time,
either.  This means the pause button must be insensitive, and we don't
want to see any warnings if we unfocus the window prior to making the
first move.