Remove game from history after "Abandon Game"
When starting a new game despite not having completed the current game,
the user may choose to either "Abandon Game" or to save the game.
However, Abandon Game currently autosaves the game so that it will be
loaded in the future.  This is annoying as the user explicitly decided
to not save the game.  Actually discard the game in this scenario.

https://bugzilla.gnome.org/show_bug.cgi?id=705878