Remove show history option
There's really no benefit to being able to hide the history bar.  You
currently gain a bit of space, but that's because this option breaks
timed games by hiding the timers.  This option might make sense after a
bit of a UI redesign, but I don't see that happening anytime soon.

https://bugzilla.gnome.org/show_bug.cgi?id=705046