Drop Amundsen support
Amundsen doesn't support the CECP Undo command.  We could work around
this in GNOME Chess, but as Amundsen is really the *only* violator that
I'm aware of, and isn't a very well-known engine, I have no inclination
to do so.  Undo is part of the core protocol and it's not an optional
item.

https://bugzilla.gnome.org/show_bug.cgi?id=705013