Remove ChessHistory class
I'm not sure why this class originally existed, but currently it's only
used to maintain a stack of autosaved games, so that after completing
the first autoloaded game on the stack, the next time you start the game
the next game on the stack will be autoloaded.  This feature was
determined to be undesirable -- we'd really like only one game to be
autoloaded at a time -- so I added a delete query to ChessHistory such
that all the previous games on the stack would be removed when a new one
was added.  But that meant an SQLite database was used to store exactly
one game.  Let's drop this.

We can also remove the dependency on SQLite.

https://bugzilla.gnome.org/show_bug.cgi?id=705878