Look for AI profiles in two different locations
We'll now search for AI profiles in:

* ~/.config/gnome-chess/engines.conf (first)
* /etc/gnome-chess/engines.conf

I'm torn as to whether to continue reading /etc/chess-engines.conf. I
think it'd be best to ignore it -- if users modified it manually to add
an engine, they will notice that the engine has disappeared. They're not
likely to notice that they're not getting the important GNUChess
improvements I'm hoping to land soon....