Do not unconditionally call prctl in distributed C
The existing check for linux/prctl.h works great if you are compiling
the Vala code, but it's expected that users and distributions may start
with the distributed C files instead. If the tarball is generated on
Linux, then the call to prctl() will unconditionally appear in the
distributed C, and if the tarball isn't generated on Linux, then it will
be unconditionally absent.

There are only two ways to fix this: (a) not distribute any C code, or
(b) move the conditional compilation out of the Vala code. Though it
looks like the Autotools Vala world may be starting to lean towards (a),
let's be conservative and pick (b) for now.

https://bugzilla.gnome.org/show_bug.cgi?id=726614