ui: remove unnecessary size groups
The horizontal size group is exacerbating the horrible resizing
behavior as the clock ticks down. It's not the size group's fault that
this is broken, but the size group makes it 1634435x worse.

The vertical size group is simply unnecessary in 3.12. In 3.10 it is
needed for the header bar buttons to have the same height, but forces
the buttons into odd (non-square) shapes. Let's just remove it.

https://bugzilla.gnome.org/show_bug.cgi?id=726951