Rephrase the player death message
For whatever reason, PGN has a notion of player death, to be used if the
moons align and a player is unable to complete a game because he died.
Probably no gnome-chess user will ever see this message ever -- he would
have to load a specially-prepared PGN -- but let's make it less
confusing anyway.

https://bugzilla.gnome.org/show_bug.cgi?id=726454