Fix various mistakes in ChessEngine
fds can be 0 (hi stdin!). Various resource leaks due to not calling
stop() when the engine stops itself. Be careful to unset all values from
the previous engine in stop(), since they're now checked to ensure
they're not overwritten as a precondition of start(). Expect
engine_stopped_cb() to be called after stop().

https://bugzilla.gnome.org/show_bug.cgi?id=724504