Fix claim draw dialog appearing for moves through history
Using the history viewer should not trigger the claim draw dialog. I
noticed this because there's a big warning comment in game_turn_cb()
saying to not use it for anything related to the game itself, yet here
it was.

Note I add one additional check here, to mave sure the game has not
ended before telling the engine to move.

Note also that this makes every move 0.1s slower than before... that
seems fine.