Fix draws claimed by engines
Currently when the engine claims a draw, we check to see if it's
possible to claim a draw. If so, we do nothing and the game stalls
forever while GNOME Chess claims the engine is thinking. If not, the
game ends in a bug (as it should).

We're missing the call to actually end the game in a draw. Fix it.

Fixes #32