Try to avoid confusion in claim draw dialog
Now that the engine can actually claim draws, it could be confusing if
the human player chooses not to claim a draw and then the engine
immediately does so. E.g. when playing against HoiChess, which always
claims a draw on threefold repetition, there is no difference between
the two options: if the human player doesn't claim a draw, the engine
will. This is hard to fix without making substantial changes to how we
present the result of the game, so let's just try to explain it instead.