Ensure the engine is always reaped
Since 8ba6bcb9929d33f273270ea0c90a97ff920aed5b, engines are only reaped
when the engine stops unexpectedly. Oops. We need to ensure the
ChildWatch callback always executes, because we rely on the
GChildWatchSource to cause GLib to reap it for us.

This ensures we don't wind up with zombie engines.