Fix layout changing when menu button is pushed.
Use the size_allocate event instead of the configure event to update the
layout on window size change. The configure event receives an event when
the menu button is pushed, which provides the size of the Popover menu,
not the ApplicationWindow. The size_allocate event does not receive this
event.