threefold/fivefold repetition checks should consider only current state
Way back in 4cabc410857685cfa2e27dc32572ee3e96480f68, I thought it was
important to allow users to claim a draw at any point in the future
after a threefold repetition. But this is not what the laws of chess
allow: it should be claimed when it happens, not at random points in the
future. At the time, that change probably made sense because we didn't
tell the user when it's possible to claim a draw, which nowadays we do
by opening a message dialog. But nowadays, it has no benefit, and a
large cost: the message dialog appears at the start of every turn,
forevermore, getting in the way and irritating the user. So let's only
show it if the current board state really is eligible for a draw.

Fixes #47


(cherry picked from commit 34b6e0dee49b0fc67b3b46129d3be2c2e0062e16)