Fix "engine error" on new game after engine moves during pause
When we replay the engine's move later, we should make sure we are still
playing the same game AND that the state of the game is still the same.
Otherwise, discard the engine's move. The user has done something to
advance the state of the game beyond what the engine knows about, so
it's totally expected that the engine's move may no longer apply.

The state of the game ought to always be the same as it was before. Add
an assert just to be sure.

Halfway fixes #49