Better handle case where window subtitle is the filename
This never looks good in narrow mode, so if we are in narrow mode, never
set the window subtitle.

But if we start in normal mode, then the user switches to narrow mode,
it looks weird for the subtitle to disappear and not come back. So let's
maintain it as the window subtitle. It should definitely not move to the
info bar, like other subtitles, since that looks weird.