Never open file chooser to /run
If we are sandboxed, then any files we save are going to be located
under /run inside the sandbox. But that's not where they are on the host
system, so we should try to avoid exposing this to the user. Sad reality
is that there is no way we can know where the file really exists on the
host, so let's just not preselect the file's location in this case.

Fixes #53