Fix check for engine timeout in untimed games
We currently give the engine 10 seconds to move, and end the game with
an engine error if it takes longer than that, to avoid the game stalling
forever if the engine breaks. But we use the chess clock to run this
check, meaning it doesn't work in untimed games. And games are untimed
by default. Oops.

Fix it by not relying on the chess clock.

Note: for stable branches, let's give it 30 seconds instead of 10, to be
conservative.

Fixes #54