Stop using header bar subtitle for game status...
...and always use the info bar instead, even when the screen is not
narrow.

The header bar subtitle will now exclusively be used to indicate the
current filename.

This simplifies the code, and avoids many strange bugs. As long as the
title is short, we can use it even in narrow mode. But the status that
we previously put into the subtitle is just too large, so it will need
the info bar. Extra status is rare, so it's OK to use the info bar even
in normal mode: it doesn't look good to have it up constantly, but it's
only used when an engine is not installed, or the game has ended, or an
en passant capture is performed, so it should be OK.

This fixes the subtitle getting messed up after using Undo when a game
has finished, but it also avoids strange state differences between
narrow and normal mode, and the complexity of switching between the two.