Use a longer message to explain en passant captures
This reverts 73a3af2e. There's no longer any need to use a short
message, since we will now always put it in the info bar.

Let's also italicize the text, as is appropriate. Sadly this means
translators will have to work with markup, because they may need to
remove the markup, or reorder the string, so it requires violating the
normal guidance to avoid markup in translatable strings. That's probably
fine; we should avoid it wherever possible, but not when it's really
needed.