Split engine support into a separate internal lib
The chess engine support currently lives in src/, which otherwise
handles only GUI functions. This is a little misplaced, because engines
are not related to the GUI. So lib/ would be a better location for it.

Except all the engine support is actually higher-level than lib/. So we
can nicely fit it in between src/ and lib/ by creating a second internal
lib, libengine.

This is only a small benefit, but it's nice to keep src/ focused on the
GUI.