window: don't try to disconnect from the clock's tick callback
This doesn't work because at the time we try to disconnect, game is a
different game than it was before, with a different clock. Oops.

There's no need to disconnect because the clock will always be destroyed
by the time we start a new game. In the unlikely event that ever changes
in the future, the worst that can happen is redrawing the clock labels
more often than necessary. So it should be safe to just never
disconnect.