Use pango to render clock text
The cairo toy text API is not really suitable for rendering more than a
single character of ASCII text. I'm not sure when this broke, but it's
currently failing to render both the ratio character that we use, and
also the left-to-right mark that's supposed to be invisible. Better
switch to something smarter, and pango is the answer.