Handle saved chess engine not being installed
When starting a new game, it doesn't matter if the chess engine saved in
gsettings is no longer installed. Just reset to default if it's not
there. It's pointless to keep printing a warning again and again about
the engine being missing.

But do continue to print the warning when loading a saved game (because
switching from one engine to another in the middle of a game is possibly
unexpected).