Fix engine name check from previous commit
The check I added in the previous commit makes it impossible to play
against a human player: when playing against a human, there is no AI
profile matching "human" and therefore the code falls back to the
default engine. Oops.

Let's also reorder the engine_level check to get it out of the
engine_name logic.