Inform engine of time to move (sort of)
Primary goal: ensure the engine knows we have a timeout that it is
required to respect.

Secondary goal: let the engine know if there is less time remaining on
its clock than the standard timeout.

For UCI, it's sufficient to just pass the clock and the standard
timeout.

For CECP, it's not possible to directly use the clock in the engine
layer, because only the higher layer knows which color is to move.

Fixes #15
Fixes #57