Fix default opponent selection in new game dialog
The opponent combo box in the new game dialog does not work properly if
the value of the setting has never been modified. It defaults to showing
Human as the selected opponent, but the actual underlying default is to
use the first selected engine. To fix this, simply add Human at the
bottom of the model instead of the top, such that the first engine is
first in the combo box model.

Fixes #97