Really fix default opponent selection in new game dialog
The preferences dialog and new game dialogs are both based on a
Preferences object abstraction, which the rest of GNOME Chess does not
use. This means it's possible for GNOME Chess to interpret the
preferences differently than its dialogs do. And we do that for default
opponent. The default settings value is an empty string, and the
documentation says it means "use first available AI opponent." But the
Preferences abstraction always uses Human opponent instead. Oops. Fix
this by introducing a default opponent and checking for it in the new
game dialog.

This is still a little fragile, in that the rest of Chess is not using
the Preferences abstraction, but that will be a problem for another day.

Fixes #97